+++ src/main/java/net/dehydration/mixin/MilkBucketItemMixin.java
@@ -1,53 +1,21 @@
 package net.dehydration.mixin;
 
-import java.util.Optional;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import org.spongepowered.asm.mixin.injection.At;
-
 import net.dehydration.DehydrationMain;
-import net.dehydration.access.ThirstManagerAccess;
 import net.dehydration.init.ConfigInit;
-import net.dehydration.init.EffectInit;
 import net.dehydration.misc.ThirstTooltipData;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.effect.StatusEffectInstance;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.MilkBucketItem;
 import net.minecraft.item.tooltip.TooltipData;
-import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+
+import java.util.Optional;
 
 @Mixin(MilkBucketItem.class)
 public abstract class MilkBucketItemMixin extends Item {
 
     public MilkBucketItemMixin(Settings settings) {
         super(settings);
-    }
-
-    @Inject(method = "finishUsing", at = @At(value = "HEAD"))
-    public void finishUsingMixin(ItemStack stack, World world, LivingEntity user, CallbackInfoReturnable<ItemStack> info) {
-        if (user instanceof PlayerEntity player) {
-
-            int thirstQuench = 0;
-            for (int i = 0; i < DehydrationMain.HYDRATION_TEMPLATES.size(); i++) {
-                if (DehydrationMain.HYDRATION_TEMPLATES.get(i).containsItem(stack.getItem())) {
-                    thirstQuench = DehydrationMain.HYDRATION_TEMPLATES.get(i).getHydration();
-                    break;
-                }
-            }
-            if (thirstQuench == 0) {
-                thirstQuench = ConfigInit.CONFIG.milk_thirst_quench;
-            }
-            ((ThirstManagerAccess) player).getThirstManager().add(thirstQuench);
-
-            if (!world.isClient() && world.random.nextFloat() >= ConfigInit.CONFIG.milk_thirst_chance) {
-                player.addStatusEffect(new StatusEffectInstance(EffectInit.THIRST, ConfigInit.CONFIG.potion_bad_thirst_duration / 2, 0, false, false, true));
-            }
-        }
     }
 
     @Override