+++ src/main/java/net/dehydration/mixin/PlayerEntityMixin.java
@@ -1,33 +1,29 @@
 package net.dehydration.mixin;
 
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.At.Shift;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import org.spongepowered.asm.mixin.injection.At;
-
-import net.dehydration.DehydrationMain;
 import net.dehydration.access.PlayerAccess;
 import net.dehydration.access.ThirstManagerAccess;
 import net.dehydration.init.ConfigInit;
-import net.dehydration.init.TagInit;
 import net.dehydration.thirst.ThirstManager;
-import net.minecraft.component.type.FoodComponent;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.HungerManager;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.world.Difficulty;
 import net.minecraft.world.GameRules;
 import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.At.Shift;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(PlayerEntity.class)
 public abstract class PlayerEntityMixin extends LivingEntity implements ThirstManagerAccess, PlayerAccess {
-    private ThirstManager thirstManager = new ThirstManager();
+    @Unique
+    private final ThirstManager thirstManager = new ThirstManager();
 
     @Override
     public ThirstManager getThirstManager() {
@@ -38,7 +34,7 @@ public abstract class PlayerEntityMixin extends LivingEntity implements ThirstMa
     protected HungerManager hungerManager = new HungerManager();
     @Shadow
     private int sleepTimer;
-
+    @Unique
     private int drinkTime = 0;
 
     public PlayerEntityMixin(EntityType<? extends LivingEntity> entityType, World world) {
@@ -92,39 +88,6 @@ public abstract class PlayerEntityMixin extends LivingEntity implements ThirstMa
             int hungerConsumption = ConfigInit.CONFIG.sleep_hunger_consumption;
             this.thirstManager.setThirstLevel(thirstLevel >= thirstConsumption ? thirstLevel - thirstConsumption : 0);
             this.hungerManager.setFoodLevel(hungerLevel >= hungerConsumption ? hungerLevel - hungerConsumption : 0);
-        }
-    }
-
-    @Inject(method = "eatFood", at = @At(value = "HEAD"))
-    private void eatFoodMixin(World world, ItemStack stack, FoodComponent foodComponent, CallbackInfoReturnable<ItemStack> info) {
-        int thirstQuench = 0;
-        if (stack.isIn(TagInit.HYDRATING_STEW)) {
-            thirstQuench = ConfigInit.CONFIG.stew_thirst_quench;
-        }
-        if (stack.isIn(TagInit.HYDRATING_FOOD)) {
-            thirstQuench = ConfigInit.CONFIG.food_thirst_quench;
-        }
-        if (stack.isIn(TagInit.HYDRATING_DRINKS)) {
-            thirstQuench = ConfigInit.CONFIG.drinks_thirst_quench;
-        }
-        if (stack.isIn(TagInit.STRONGER_HYDRATING_STEW)) {
-            thirstQuench = ConfigInit.CONFIG.stronger_stew_thirst_quench;
-        }
-        if (stack.isIn(TagInit.STRONGER_HYDRATING_FOOD)) {
-            thirstQuench = ConfigInit.CONFIG.stronger_food_thirst_quench;
-        }
-        if (stack.isIn(TagInit.STRONGER_HYDRATING_DRINKS)) {
-            thirstQuench = ConfigInit.CONFIG.stronger_drinks_thirst_quench;
-        }
-
-        for (int i = 0; i < DehydrationMain.HYDRATION_TEMPLATES.size(); i++) {
-            if (DehydrationMain.HYDRATION_TEMPLATES.get(i).containsItem(stack.getItem())) {
-                thirstQuench = DehydrationMain.HYDRATION_TEMPLATES.get(i).getHydration();
-                break;
-            }
-        }
-        if (thirstQuench > 0) {
-            ((ThirstManagerAccess) (PlayerEntity) (Object) this).getThirstManager().add(thirstQuench);
         }
     }