+++ src/main/java/net/dehydration/mixin/PotionItemMixin.java
@@ -1,32 +1,14 @@
 package net.dehydration.mixin;
 
-import java.util.Optional;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import org.spongepowered.asm.mixin.injection.At;
-
 import net.dehydration.DehydrationMain;
-import net.dehydration.access.ThirstManagerAccess;
 import net.dehydration.init.ConfigInit;
-import net.dehydration.init.EffectInit;
 import net.dehydration.misc.ThirstTooltipData;
-import net.dehydration.thirst.ThirstManager;
+import net.dehydration.util.ThirstHelper;
 import net.minecraft.component.DataComponentTypes;
 import net.minecraft.component.type.PotionContentsComponent;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.item.PotionItem;
-import net.minecraft.item.ThrowablePotionItem;
+import net.minecraft.item.*;
 import net.minecraft.item.tooltip.TooltipData;
-import net.minecraft.potion.Potion;
-import net.minecraft.potion.Potions;
-import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.registry.tag.FluidTags;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
@@ -37,6 +19,12 @@ import net.minecraft.util.hit.HitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.RaycastContext;
 import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.Optional;
 
 @Mixin(PotionItem.class)
 public abstract class PotionItemMixin extends Item {
@@ -57,33 +45,6 @@ public abstract class PotionItemMixin extends Item {
         }
     }
 
-    @Inject(method = "finishUsing", at = @At(value = "HEAD"))
-    public void finishUsingMixin(ItemStack stack, World world, LivingEntity user, CallbackInfoReturnable<ItemStack> info) {
-        if (user instanceof PlayerEntity player) {
-            PotionContentsComponent potionContentsComponent = stack.getOrDefault(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent.DEFAULT);
-            if (!world.isClient() && this.isBadPotion(potionContentsComponent.potion().get()) && world.random.nextFloat() >= ConfigInit.CONFIG.potion_bad_thirst_chance) {
-                player.addStatusEffect(new StatusEffectInstance(EffectInit.THIRST, ConfigInit.CONFIG.potion_bad_thirst_duration, 0, false, false, true));
-            }
-            ThirstManager thirstManager = ((ThirstManagerAccess) player).getThirstManager();
-            int thirstQuench = 0;
-            for (int i = 0; i < DehydrationMain.HYDRATION_TEMPLATES.size(); i++) {
-                if (DehydrationMain.HYDRATION_TEMPLATES.get(i).containsItem(stack.getItem())) {
-                    thirstQuench = DehydrationMain.HYDRATION_TEMPLATES.get(i).getHydration();
-                    break;
-                }
-            }
-            if (thirstQuench == 0) {
-                thirstQuench = ConfigInit.CONFIG.potion_thirst_quench;
-            }
-            thirstManager.add(thirstQuench);
-        }
-    }
-
-    private boolean isBadPotion(RegistryEntry<Potion> potion) {
-        return potion == Potions.WATER || potion == Potions.AWKWARD || potion == Potions.THICK || potion == Potions.HARMING || potion == Potions.LONG_POISON || potion == Potions.LONG_SLOWNESS
-                || potion == Potions.LONG_WEAKNESS || potion == Potions.MUNDANE || potion == Potions.POISON || potion == Potions.SLOWNESS || potion == Potions.STRONG_HARMING
-                || potion == Potions.STRONG_POISON || potion == Potions.STRONG_SLOWNESS || potion == Potions.WEAKNESS;
-    }
 
     @Override
     public Optional<TooltipData> getTooltipData(ItemStack stack) {
@@ -99,7 +60,7 @@ public abstract class PotionItemMixin extends Item {
             if (thirstQuench == 0) {
                 thirstQuench = ConfigInit.CONFIG.potion_thirst_quench;
             }
-            if (stack.getOrDefault(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent.DEFAULT).potion().isPresent() && isBadPotion(stack.getOrDefault(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent.DEFAULT).potion().get())) {
+            if (stack.getOrDefault(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent.DEFAULT).potion().isPresent() && ThirstHelper.isBadPotion(stack.getOrDefault(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent.DEFAULT).potion().get())) {
                 return Optional.of(new ThirstTooltipData(2, thirstQuench));
             }
             return Optional.of(new ThirstTooltipData(0, thirstQuench));